+++ CHANGELOG.md
@@ -6,4 +6,5 @@
 
 ## [2025-10-03]
 
--   Changed from using nx cli to list the projects to using the vscode api to find the projects ourself
+-   Changed from nx cli to list projects to using the vscode api to find the projects ourself
+-   Changed from nx cli to list targets, to reading $.targets from the project.json of the selected project